By using add a new sacffoled item i have add and identity page where user and sign in and login.
I have changed the fonts on the layout page and i have use the website google fonts and cdnjs.com. Under the wwwroot family I changed the font family to Roboto and sans-serif.

Under the identity folder then data then user i had to put and public string for first and last name after i had to connect the login and register page to sql server using the appsetting and then using Nuget package manager-  package manager console.

Add a new razor layoout page and add some div classes to the page and connected the page to the layout page for login and register page.

On the Register page I modified the page by adding the first name and last name.. I had to make sure when entering a password that its not using a deafult password requirement from ASP NET. and to shoow the login name i had to add a partial name on the deafault Layout page