I had issues with the About page, to fix this i had go to the home controllers page and change the index page to the About pages. I have added tabs inside the About page so i dont have to make 2 different razor pages.
History- there a old picture of avondale and a brief of avondale college

Motto and Mission Statement- on this page there a picture of the school motto and the school mission/ goals and Avondale college heart values.

Message from the Principal- there is a picture Mrs Watkinson the principal of avondale college and a message for the viewer.